lots of permissions and constants work

+++ api-v1/paths/deaddrop/{deaddrop_id}.js
@@ -9,19 +9,27 @@ module.exports = function(toolKit, deaddropService, messageService, permissionSe
 		try {
 			var user_id = req.session.user_id;
 			var deaddrop_id = req.params.deaddrop_id;
+			
+			console.log(req.session);
 
-			deaddropService.getDeadDrop(user_id, deaddrop_id)
-				.then((deaddropObj) => {
-					if (deaddropObj) {
-						messageService.getDeadDropMessages(user_id, deaddrop_id)
-.then( (messages) => {
-							deaddropObj.messages = messages;
-							res.status(200).json(deaddropObj);
-						});
-					} else {
-						res.status(404).json(toolKit.createSimpleResponse("error", "deaddrop not found:" + deaddrop_id));
-					}
-				});
+			//verify the user has permission to create a deaddrop
+			if (user2PermissionService.isCacheUserPermission(req, deaddrop_id,toolKit.getConstants().SYS_DETAILS_READ)) {
+
+				deaddropService.getDeadDrop(user_id, deaddrop_id)
+					.then((deaddropObj) => {
+						if (deaddropObj) {
+							messageService.getDeadDropMessages(user_id, deaddrop_id)
+								.then((messages) => {
+									deaddropObj.messages = messages;
+									res.status(200).json(deaddropObj);
+								});
+						} else {
+							res.status(404).json(toolKit.createSimpleResponse("error", "deaddrop not found:" + deaddrop_id));
+						}
+					});
+			} else {
+				res.status(401).json(toolKit.createSimpleResponse("error", `${user_id} does not have ${toolKit.getConstants().SYS_DETAILS_READ} permissions for ${deaddrop_id}`));
+			}
 		}
 		catch (e) {
 			res.status(500).json(toolKit.createSimpleResponse("error", e.message));
@@ -60,35 +68,40 @@ module.exports = function(toolKit, deaddropService, messageService, permissionSe
 			var deaddrop_id = req.params.deaddrop_id;
 			var deaddrop_key = req.body.deaddrop_key;
 			var title = req.body.title;
-			var details = "CREATE READ UPDATE DELETE";
 			var tags = "DEADDROP";
 
-			permissionService.getPermission(deaddrop_id)
-				.then(function(permissionObj) {
-					if (permissionObj) {
-						res.status(401).json(toolKit.createSimpleResponse("error", "This deaddrop name is alread taken"));
-					} else {
-						permissionService.createPermission(deaddrop_id, deaddrop_id, tags)
-							.then(() => {
-								user2PermissionService.addUserPermission(user_id, deaddrop_id, details)
-									.then(() => {
-										deaddropService.createNewDeadDrop(deaddrop_id, title, deaddrop_key)
-											.then(() => {
-												messageObj = {
-													"deaddrop_id": deaddrop_id,
-													"user_id": user_id,
-													"title": "Welcome to " + deaddrop_id,
-													"message": "This is " + deaddrop_id
-												}
-												messageService.addMessage(messageObj)
-.then( () => {
-													res.status(200).json(toolKit.createSimpleResponse("success", "deaddrop added"));
-												});
-											})
-									})
-							})
-					}
-				});
+			if (user2PermissionService.isCacheUserPermission(req,  toolKit.getConstants().DEADDROP_ADMIN,toolKit.getConstants().SYS_DETAILS_CREATE)) {
+
+				// if the permission exists than the deaddrop is already taken
+				permissionService.getPermission(deaddrop_id)
+					.then(function(aPermissionObj) {
+						if (aPermissionObj) {
+							res.status(401).json(toolKit.createSimpleResponse("error", "This deaddrop name is already taken"));
+						} else {
+							permissionService.createPermission(deaddrop_id, deaddrop_id, tags, deaddrop_key)
+								.then(() => {
+									user2PermissionService.addUserPermission(user_id, deaddrop_id, toolKit.getConstants().SYS_DETAILS_ALL)
+										.then(() => {
+											deaddropService.createNewDeadDrop(deaddrop_id, title)
+												.then(() => {
+													messageObj = {
+														"deaddrop_id": deaddrop_id,
+														"user_id": user_id,
+														"title": "Welcome to " + deaddrop_id,
+														"message": "This is " + deaddrop_id
+													}
+													messageService.addMessage(messageObj)
+														.then(() => {
+															res.status(200).json(toolKit.createSimpleResponse("success", "deaddrop added"));
+														});
+												})
+										})
+								})
+						}
+					});
+			} else {
+				res.status(401).json(toolKit.createSimpleResponse("error", `${user_id} does not have ${toolKit.getConstants().SYS_DETAILS_CREATE} permissions for ${deaddrop_id}`));
+			}
 		}
 		catch (e) {
 			res.status(500).json(toolKit.createSimpleResponse("error", e.message));
@@ -136,11 +149,22 @@ module.exports = function(toolKit, deaddropService, messageService, permissionSe
 		try {
 			var user_id = req.session.user_id;
 			var deaddrop_id = req.params.deaddrop_id;
-			console.log('DELETE /deaddrop/', deaddrop_id);
-			deaddropService.deleteDeadDrop(user_id, deaddrop_id)
-				.then(() => {
-					res.status(200).json(toolKit.createSimpleResponse("success", "deaddrop removed"));
-				})
+
+			if (user2PermissionService.isCacheUserPermission(req,  deaddrop_id,toolKit.getConstants().SYS_DETAILS_DELETE)) {
+				user2PermissionService.deleteUserPermissionsByPermission(deaddrop_id)
+					.then(() => {
+						res.stats(200).json(toolKit.createSimpleResponse("success", "deaddrop removed"));
+					})
+				messageService.deleteDeadDropMessages(deaddrop_id)
+					.then(() => {
+						deaddropService.deleteDeadDrop(user_id, deaddrop_id)
+							.then(() => {
+								permissionService.deletePermission(deaddrop_id)
+							})
+					})
+			} else {
+				res.status(401).json(toolKit.createSimpleResponse("error", `${user_id} does not have ${toolKit.getConstants().SYS_DETAILS_DELETE} permissions for ${deaddrop_id}`));
+			}
 		}
 		catch (e) {
 			res.status(500).json(toolKit.createSimpleResponse("error", e.message));
@@ -148,7 +172,7 @@ module.exports = function(toolKit, deaddropService, messageService, permissionSe
 	};
 
 	DELETE.apiDoc = {
-		"summary": "Delete a deaddrop connection",
+		"summary": "Delete a deaddrop",
 		"parameters": [
 			{
 				"in": 'path',
